Campos Virtuales Modelos

+++ ERPAPI/Models/UserToken.cs
@@ -12,7 +12,7 @@ namespace ERPAPI.Models
 
         public DateTime Expiration { get; set; }
 
-        public Int64 BranchId { get; set; }
+        public int BranchId { get; set; }
 
         public DateTime LastPasswordChangedDate { get; set; }
